Make CmdCp easier to use
The name suggest running a command and copy files.
However the commands need to be specified as a regex-map list which
might be executed for every source instead of once.
Often a workflow might be just `./custom_configure && ./custom_build.sh`
which is not easy to map to this.
Instead allow non-tuples which are considered as commands to be run.
When mixing tuples and non-tuples the plain commands are run last.