Add `build_target` parameter to `PythonPackage`
Using `buildcmd` to pass options to `setup.py` is confusing and makes it
impossible to fully replace the build command.
Introduce `build_target` similar to `install_target` and deprecate
passing a target and optional options via `buildcmd`.

The old behavior is used when `buildcmd` is a known setup.py command
such as `build` or `build_ext` or a single word, e.g. `clean`.